+++ packages/frontend-shared/cypress/support/mock-graphql/mountFragment.ts
@@ -30,6 +30,7 @@ export const registerMountFn = ({ plugins }: MountFnOptions = {}) => {
       options.global.stubs = options.global.stubs || {}
       if (!Array.isArray(options.global.stubs)) {
         options.global.stubs.transition = false
+        options.global.stubs['transition-group'] = false
       }
 
       options.global.plugins = options.global.plugins || []
@@ -51,7 +52,7 @@ export const registerMountFn = ({ plugins }: MountFnOptions = {}) => {
     },
   )
 
-  function mountFragment<T extends TypedDocumentNode<any, any>> (source: T, options: MountFragmentConfig<T>, list: boolean = false): Cypress.Chainable<ClientTestContext> {
+  function mountFragment<T extends TypedDocumentNode<any, any>> (source: T, options: MountFragmentConfig<T>, list: boolean = false): Cypress.Chainable<any> {
     let hasMounted = false
     const context = makeClientTestContext()
     const fieldName = list ? 'testFragmentMemberList' : 'testFragmentMember'
@@ -60,6 +61,7 @@ export const registerMountFn = ({ plugins }: MountFnOptions = {}) => {
       global: {
         stubs: {
           transition: false,
+          'transition-group': false,
         },
         plugins: [
           createI18n(),
@@ -131,7 +133,7 @@ export const registerMountFn = ({ plugins }: MountFnOptions = {}) => {
 
         return props.gql ? options.render(props.gql) : h('div')
       },
-    }), mountingOptions).then(() => context)
+    }), mountingOptions)
   }
 
   const mutationResolvers: Map<string, MutationResolver<any>> = new Map()
@@ -164,6 +166,11 @@ export const registerMountFn = ({ plugins }: MountFnOptions = {}) => {
     Cypress.log({
       name: 'subscription event',
       message: name,
+      consoleProps: () => {
+        return {
+          document,
+        }
+      },
     })
 
     const hook = subscriptionHooks.get(name)
@@ -245,7 +252,7 @@ declare global {
       mountFragment<T extends TypedDocumentNode<any, any>>(
         fragment: T,
         config: MountFragmentConfig<T>
-      ): Cypress.Chainable<ClientTestContext>
+      ): Cypress.Chainable<any>
 
       /**
        * mock a mutation resolver when needed to spy on it or modify the result
@@ -262,7 +269,7 @@ declare global {
       mountFragmentList<T extends TypedDocumentNode<any, any>>(
         fragment: T,
         config: MountFragmentListConfig<T>
-      ): Cypress.Chainable<ClientTestContext>
+      ): Cypress.Chainable<any>
 
       /**
        * Helper to register an event to be returned when testing subscriptions
@@ -270,25 +277,17 @@ declare global {
        * @param emitEvent callback that you must call with the document result to pass to the subscription event
        *
        * Example:
-       *
        * cy.stubSubscriptionEvent(DebugPendingRunSplash_SpecsDocument, () => {
-       *  return {
-       *    relevantRunSpecChange: {
-       *      __typename: 'Query',
-       *      currentProject: {
-       *        __typename: 'CurrentProject',
-       *        id: 'fake',
-       *        relevantRunSpecs: {
-       *          __typename: 'CurrentProjectRelevantRunSpecs',
-       *          current: {
-       *            __typename: 'RelevantRunSpecs',
-       *            completedSpecs: completed,
-       *            totalSpecs: total,
-       *          },
-       *        },
-       *      },
-       *    },
-       *  }å
+       *   return {
+       *     __typename: 'Subscription' as const,
+       *     cloudNode: {
+       *       __typename: 'CloudRun' as const,
+       *       id: 'fake',
+       *       totalSpecs: total,
+       *       completedSpecs: completed,
+       *       scheduledToCompleteAt,
+       *     },
+       *   }
        * })
        *
        */