+++ scripts/binary.js
@@ -1,4 +1,3 @@
-/* eslint-disable no-console */
 require('@packages/ts/register')
 
 const command = process.argv[2]